third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-svg.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>

description("Test createImageBitmap from a SVG image.");
window.jsTestIsAsync = true;

function shouldBeGreen(x, y) {
    d = ctx.getImageData(x, y, 1, 1).data;
    shouldBeEqualToNumber("d[0]", 0);
    shouldBeEqualToNumber("d[1]", 128);
    shouldBeEqualToNumber("d[2]", 0);
    shouldBeEqualToNumber("d[3]", 255);
}

function shouldBeTransparent(x, y) {
    d = ctx.getImageData(x, y, 1, 1).data;
    shouldBeEqualToNumber("d[0]", 0);
    shouldBeEqualToNumber("d[1]", 0);
    shouldBeEqualToNumber("d[2]", 0);
    shouldBeEqualToNumber("d[3]", 0);
}

function checkBitmap(imageBitmap, width, height)
{
    bitmap = imageBitmap;
    if (width == 200) {
        shouldBe("bitmap.width", "200");
        shouldBe("bitmap.height", "200");
    } else if (width == 100) {
        shouldBe("bitmap.width", "100");
        shouldBe("bitmap.height", "100");
    } else if (width == 50) {
        shouldBe("bitmap.width", "50");
        shouldBe("bitmap.height", "50");
    }
    var half_width = width / 2;
    var half_height = height / 2;
    ctx.clearRect(0, 0, 200, 200);
    ctx.drawImage(bitmap, 0, 0);
    shouldBeGreen(0, 0);
    shouldBeGreen(0, half_height - 1);
    shouldBeGreen(half_width - 1, 0);
    shouldBeGreen(half_width - 1, half_height - 1);
    shouldBeTransparent(0, height - 1);
    shouldBeTransparent(width - 1, 0);
    shouldBeTransparent(width - 1, height - 1);
}

function checkFlippedBitmap(imageBitmap, width, height)
{
    bitmap = imageBitmap;
    if (width == 200) {
        shouldBe("bitmap.width", "200");
        shouldBe("bitmap.height", "200");
    } else if (width == 100) {
        shouldBe("bitmap.width", "100");
        shouldBe("bitmap.height", "100");
    } else if (width == 50) {
        shouldBe("bitmap.width", "50");
        shouldBe("bitmap.height", "50");
    }
    var half_width = width / 2;
    var half_height = height / 2;
    ctx.clearRect(0, 0, 200, 200);
    ctx.drawImage(bitmap, 0, 0);
    shouldBeTransparent(0, 0);
    shouldBeTransparent(0, half_height - 1);
    shouldBeTransparent(half_width - 1, 0);
    shouldBeTransparent(half_width - 1, half_height - 1);
    shouldBeGreen(0, half_height);
    shouldBeGreen(half_width - 1, half_height);
    shouldBeGreen(0, height - 1);
    shouldBeGreen(half_width - 1, height - 1);
}

var canvas = document.createElement("canvas");
canvas.setAttribute("width", "200");
canvas.setAttribute("height", "200");
var ctx = canvas.getContext("2d");
var bitmap;

var image = new Image();
image.onload = function() {
    imageBitmaps = {};
    var p1 = createImageBitmap(image).then(function(image) { imageBitmaps.noResize = image });
    var p2 = createImageBitmap(image, {resizeWidth: 50, resizeHeight: 50, resizeQuality: "high"}).then(function(image) { imageBitmaps.resizeDown = image });
    var p3 = createImageBitmap(image, 50, 50, 100, 100).then(function(image) { imageBitmaps.cropCenter = image });
    var p4 = createImageBitmap(image, 50, 50, 100, 100, {resizeWidth: 200, resizeHeight: 200, resizeQuality: "high"}).then(function(image) { imageBitmaps.cropResizeUp = image });
    var p5 = createImageBitmap(image, 50, 50, 100, 100, {resizeWidth: 50, resizeHeight: 50, resizeQuality: "high"}).then(function(image) { imageBitmaps.cropResizeDown = image });
    var p6 = createImageBitmap(image, {imageOrientation: "flipY"}).then(function(image) { imageBitmaps.noResizeFlip = image });
    var p7 = createImageBitmap(image, 50, 50, 100, 100, {imageOrientation: "flipY"}).then(function(image) { imageBitmaps.cropCenterFlip = image });
    var p8 = createImageBitmap(image, {resizeWidth: 100, resizeHeight: 100, resizeQuality: "high", imageOrientation: "flipY"}).then(function(image) { imageBitmaps.resizeFlip = image });
    var p9 = createImageBitmap(image, 50, 50, 100, 100, {resizeWidth: 200, resizeHeight: 200, resizeQuality: "high", imageOrientation: "flipY"}).then(function(image) { imageBitmaps.resizeFlip = image });
    Promise.all([p1, p2, p3, p4, p5, p6, p7, p8, p9]).then(function() {
        checkBitmap(imageBitmaps.noResize, 200, 200);
        checkBitmap(imageBitmaps.resizeDown, 50, 50);
        checkBitmap(imageBitmaps.cropCenter, 100, 100);
        checkBitmap(imageBitmaps.cropResizeUp, 200, 200);
        checkBitmap(imageBitmaps.cropResizeDown, 50, 50);
        checkFlippedBitmap(imageBitmaps.noResizeFlip, 200, 200);
        checkFlippedBitmap(imageBitmaps.cropCenterFlip, 100, 100);
        checkFlippedBitmap(imageBitmaps.resizeFlip, 100, 100);
        checkFlippedBitmap(imageBitmaps.resizeFlip, 200, 200);
        finishJSTest();
    }, function (e) {
        testFailed("createImageBitmap promise rejected: " + e);
        finishJSTest();
    });
}
image.src = '../../svg/as-image/resources/100px-green-rect.svg';

</script>
</body>
</html>
