third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-svg-expected.txt

Test createImageBitmap from a SVG image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bitmap.width is 200
PASS bitmap.height is 200
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS bitmap.width is 50
PASS bitmap.height is 50
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS bitmap.width is 100
PASS bitmap.height is 100
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS bitmap.width is 200
PASS bitmap.height is 200
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS bitmap.width is 50
PASS bitmap.height is 50
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS bitmap.width is 200
PASS bitmap.height is 200
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS bitmap.width is 100
PASS bitmap.height is 100
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
FAIL bitmap.width should be 100. Was 200.
FAIL bitmap.height should be 100. Was 200.
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 128
PASS d[2] is 0
PASS d[3] is 255
PASS bitmap.width is 200
PASS bitmap.height is 200
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
FAIL d[1] should be 0. Was 128.
PASS d[2] is 0
FAIL d[3] should be 0. Was 255.
PASS d[0] is 0
FAIL d[1] should be 128. Was 0.
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
PASS d[0] is 0
FAIL d[1] should be 128. Was 0.
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
PASS d[0] is 0
FAIL d[1] should be 128. Was 0.
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
PASS d[0] is 0
FAIL d[1] should be 128. Was 0.
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
